Add floating usm priority

+++ flyins/models.py
@@ -13,7 +13,7 @@ class FlyIn(models.Model):
     last_names = models.CharField(max_length=128, verbose_name='Apellidos')
 
     rol = models.CharField(max_length=20, verbose_name='Rol USM')
-    usm_priority = models.IntegerField(verbose_name='Prioridad')
+    usm_priority = models.FloatField(verbose_name='Prioridad')
     parallel = models.IntegerField()
 
     previous_experience = models.BooleanField()
